refactor: :wrench: share config between plugins
